Revert "Merge remote-tracking branch 'upstream/develop' into develop-mac"
This reverts commit 482bc4d5aa1a6542724d530b5ffc2f604bc0f95d, reversing
changes made to 226c4bc1c4b0ba9234946345dd7e3fe4ed0a6766.

+++ hi_scripting/scripting/components/ScriptingContentComponent.cpp
@@ -563,29 +563,34 @@ void ScriptContentComponent::makeScreenshot(const File& target, Rectangle<float>
 {
 	WeakReference<ScriptContentComponent> safeThis(this);
 
-	SafeAsyncCall::callAsyncIfNotOnMessageThread<ScriptContentComponent>(*this, [target, area](ScriptContentComponent& st)
+	auto f = [safeThis, target, area]()
 	{
-		ScriptingObjects::ScriptShader::ScopedScreenshotRenderer ssr;
-
-		auto sf = UnblurryGraphics::getScaleFactorForComponent(&st);
-
-		st.repaint();
-
-		auto img = st.createComponentSnapshot(area.toNearestInt(), true, sf);
-
-		juce::PNGImageFormat png;
-
-		target.deleteFile();
-
-		FileOutputStream fos(target);
-
-		auto ok = png.writeImageToStream(img, fos);
-
-		if (ok)
+		if (safeThis != nullptr)
 		{
-			debugToConsole(dynamic_cast<Processor*>(st.processor), "Screenshot exported as " + target.getFullPathName());
+			ScriptingObjects::ScriptShader::ScopedScreenshotRenderer ssr;
+
+			auto sf = UnblurryGraphics::getScaleFactorForComponent(safeThis.get());
+
+			safeThis->repaint();
+
+			auto img = safeThis->createComponentSnapshot(area.toNearestInt(), true, sf);
+
+			juce::PNGImageFormat png;
+
+			target.deleteFile();
+
+			FileOutputStream fos(target);
+
+			auto ok = png.writeImageToStream(img, fos);
+
+			if (ok)
+			{
+				debugToConsole(dynamic_cast<Processor*>(safeThis->processor), "Screenshot exported as " + target.getFullPathName());
+			}
 		}
-	});
+	};
+
+	MessageManager::callAsync(f);
 }
 
 void ScriptContentComponent::visualGuidesChanged()
